requirements.txt
Flask
gunicorn
numpy
pillow
altair
pandas
torch
torchvision
open-clip-torch
opencv-python